backend/requirements.txt
fastapi
uvicorn
pandas
google-generativeai
python-multipart
requests
beautifulsoup4
yfinance
lightweight-charts
plotly
python-dotenv
GoogleNews
pywebview
